marketing_analysis/requirements.txt
asgiref==3.8.1
dj-config-url==0.1.1
Django==4.2.15
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary==2.9.9
pandas==2.2.2

